Add CoinHuntingPageInner component: Refactor CoinHuntingPage to utilize CoinHuntingPageInner for improved modularity. Update modals and state management for better user experience. Standardize 'use client' directive across components.
