+++ frontend/student-web/app/(app)/community/coin-hunting/_components/CoinHuntingHelpers.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client"
 
 import React, { useEffect } from 'react';
 import { animated, useSpring } from '@react-spring/web';
