+++ frontend/student-web/app/(app)/community/coin-hunting/_components/InteractiveMap.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client"
 
 import React, { useRef, useState, useEffect, useCallback } from 'react';
 import { useGesture } from '@use-gesture/react';
