+++ frontend/student-web/app/(app)/community/coin-hunting/_components/MarkerDetailModal.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client"
 
 import React from 'react';
 import Image from 'next/image';
