+++ frontend/student-web/app/(app)/community/coin-hunting/_components/StampModal.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React, { useEffect, useRef, useState } from 'react';
 import { motion, AnimatePresence, useAnimation } from 'framer-motion';
