+++ frontend/student-web/app/(app)/community/coin-hunting/_components/TopBar.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client"
 
 import React from 'react';
 import { useRouter } from 'next/navigation';
